requirements.txt
aioredis==1.3.1
amqp==2.6.1
apipkg==1.5
APScheduler==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.12.3
Automat==20.2.0
billiard==3.5.0.5
celery==4.2.0
certifi==2020.6.20
cffi==1.14.4
channels==3.0.3
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
confluent-kafka==1.5.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==3.3.1
daphne==3.0.1
DateTime==4.3
Django==3.0.7
django-ckeditor==5.4.0
django-cors-headers==3.6.0
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==0.1.1
django-model-utils==3.1.2
django-smtp-ssl==1.0
djangorestframework==3.12.2
docxtpl==0.6.3
drf-extensions==0.6.0
drf-yasg==1.20.0
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
et-xmlfile==1.0.1
execnet==1.7.1
factory-boy==2.11.1
Faker==4.1.3
geographiclib==1.50
gunicorn==19.7.1
h11==0.12.0
hiredis==1.1.0
httptools==0.1.1
hyperlink==20.0.1
idna==2.6
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.1
IP2Location==8.5.1
IPy==1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kafka-python==2.0.1
kombu==4.6.11
lml==0.0.9
lxml==4.5.2
MarkupSafe==1.1.1
memory-profiler==0.57.0
msgpack==0.6.2
netifaces==0.10.7
numpy==1.19.4
olefile==0.46
opcua==0.98.12
openpyxl==3.0.5
packaging==19.1
pandas==1.1.5
Pillow==4.2.1
ply==3.11
pprofile==2.0.5
prompt-toolkit==3.0.8
psutil==5.7.3
psycopg2==2.7.3.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.8
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.8
PyHamcrest==2.0.2
pymodbus==2.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
pytest==3.2.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-env==0.6.2
pytest-forked==0.2
pytest-xdist==1.22.0
python-dateutil==2.8.1
python-docx==0.8.7
python-snap7==0.11
pytz==2020.1
PyYAML==5.3
redis==3.5.3
regex==2020.10.15
reportlab==3.5.42
requests==2.18.4
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.22
uvicorn==0.13.3
uvloop==0.14.0
uWSGI==2.0.19.1
vine==1.3.0
wcwidth==0.2.5
websockets==8.1
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.1.2
